Transfer testing
Code was written to not space balls out in transfer
There is comment in the controller explaining the code... but the code
is very scuffed. The comment also uses wrong names as bii should be ball
in *transfer* not intake but ill fix that later.
Ball counts were not getting properly reset as they only got reset on
first call of Teleop. This causes an issue as going from diabled to
transfer and back again would not cause a reset, so there is now code in
Robot Main to properly reset
A few new globals were added: shooting, ball in transfer, ball in flang